requirements.txt
pydantic
azure-cosmos
python-dotenv
azure-identity
azure-functions